requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
